pyproject.toml
[project]
name = "secure-exam-portal"
version = "1.0.0"
description = "A secure online exam portal built with Flask for MCA projects"
authors = [
    { name = "Sameer Alam", email = "sameeralam3127@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    "Flask==3.0.2",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Login==0.6.3",
    "Flask-WTF==1.2.1",
    "Werkzeug==3.0.6",
    "python-dotenv==1.0.1",
    "SQLAlchemy==2.0.41",
    "ruff>=0.14.7",
    "black>=25.11.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["app"]
